forms: move ArXivField with the other fields
